[FEATURE] Add `CreditDebitIdentifier` to entry transaction detail
The credit-debit-identifier (CrdDbtInd) have been used for
several details, e.g. amount, to proper represent values.
The information itself is not added.

As it could not be reliable retrieved from subvalues what
the original state was, this change now adds that detail
directly to the transaction details and the entry itself.

Test fixture files are updated to contain the added fields.

Resolves: #146
Related: #116

Conflicts:
	src/DTO/Entry.php
	src/DTO/EntryTransactionDetail.php
	src/Decoder/EntryTransactionDetail.php